GAMEPLAY: The concept of 'scope' no longer applies, and data and reliability is universal.  This is a major change that paves the way for newer features coming soon.

+++ TestFlightCore/TestFlightCore/UserSettings.cs
@@ -12,13 +12,13 @@ namespace TestFlightCore
 
 
         [Persistent] public bool debugLog = false;
-        [Persistent] public double minTimeBetweenDataPoll = 0.5;
-        [Persistent] public double minTimeBetweenFailurePoll = 60;
+        [Persistent] public float minTimeBetweenDataPoll = 0.5f;
+        [Persistent] public float minTimeBetweenFailurePoll = 60;
         [Persistent] public bool processAllVessels = false;
-        [Persistent] public double flightDataMultiplier = 1.0;
-        [Persistent] public double flightDataEngineerMultiplier = 1.0;
-        [Persistent] public double globalReliabilityModifier = 1.0;
-        [Persistent] public double masterStatusUpdateFrequency = 0.25;
+        [Persistent] public float flightDataMultiplier = 1.0f;
+        [Persistent] public float flightDataEngineerMultiplier = 1.0f;
+        [Persistent] public float globalReliabilityModifier = 1.0f;
+        [Persistent] public float masterStatusUpdateFrequency = 0.25f;
         [Persistent] public bool displaySettingsWindow = false;
         [Persistent] public int settingsPage = 0;
         [Persistent] public bool enableHUD = false;
